+++ gridam/src/features/auth/login/api/login-action.ts
@@ -1,15 +1,12 @@
-export type LoginResult = { ok: boolean; message: string }
+import { redirect } from 'next/navigation'
 
-export async function loginAction(
-  _prev: LoginResult | null,
-  formData: FormData
-): Promise<LoginResult> {
+export async function loginAction(formData: FormData): Promise<void> {
   const email = String(formData.get('email') ?? '')
   const password = String(formData.get('password') ?? '')
 
-  if (!email && !password) return { ok: false, message: '이메일과 비밀번호를 입력하세요.' }
-  if (!email) return { ok: false, message: '이메일을 입력하세요.' }
-  if (!password) return { ok: false, message: '비밀번호가 올바르지 않습니다.' }
+  if (!email && !password) redirect('/login?message=이메일과 비밀번호를 입력하세요.')
+  if (!email) redirect('/login?message=이메일을 입력하세요.')
+  if (!password) redirect('/login?message=비밀번호가 올바르지 않습니다.')
 
-  return { ok: true, message: '로그인 성공' }
+  redirect('/?message=로그인 성공')
 }
